Move RealKey to Extensions class
